Add DeleteClientCommandParser
