+++ src/test/java/seedu/address/logic/parser/DeleteCommandParserTest.java
@@ -7,6 +7,7 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 
 import org.junit.Test;
 
+import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.DeleteCommand;
 
 /**
@@ -18,11 +19,11 @@ import seedu.address.logic.commands.DeleteCommand;
  */
 public class DeleteCommandParserTest {
 
-    private DeleteCommandParser parser = new DeleteCommandParser();
+    private DeleteCommandParser parser = new DeleteClientCommandParser();
 
     @Test
     public void parse_validArgs_returnsDeleteCommand() {
-        assertParseSuccess(parser, "1", new DeleteCommand(INDEX_FIRST_PERSON));
+        assertParseSuccess(parser, "1", new DeleteCommand(INDEX_FIRST_PERSON, Command.ContactType.CLIENT));
     }
 
     @Test
